Made Serdes injectable.  Added plumbing to enforce table filter templates.  Added tests
