+++ elide-datastore/elide-datastore-aggregation/src/main/java/com/yahoo/elide/datastores/aggregation/filter/visitor/MatchesTemplateVisitor.java
@@ -5,7 +5,6 @@
  */
 package com.yahoo.elide.datastores.aggregation.filter.visitor;
 
-import com.google.common.base.Preconditions;
 import com.yahoo.elide.core.filter.FilterPredicate;
 import com.yahoo.elide.core.filter.expression.AndFilterExpression;
 import com.yahoo.elide.core.filter.expression.FilterExpression;
@@ -13,6 +12,9 @@ import com.yahoo.elide.core.filter.expression.FilterExpressionVisitor;
 import com.yahoo.elide.core.filter.expression.NotFilterExpression;
 import com.yahoo.elide.core.filter.expression.OrFilterExpression;
 import com.yahoo.elide.parsers.expression.FilterExpressionNormalizationVisitor;
+
+import com.google.common.base.Preconditions;
+
 import lombok.AllArgsConstructor;
 
 /**
